+++ packages/node/test/integration/plugins/operator/OperatorPlugin.test.ts
@@ -5,12 +5,12 @@ import {
     _operatorContractUtils
 } from '@streamr/sdk'
 import { fastPrivateKey, fetchPrivateKeyWithGas } from '@streamr/test-utils'
-import { collect, waitForCondition, StreamPartIDUtils, EthereumAddress, toEthereumAddress } from '@streamr/utils'
+import { EthereumAddress, StreamPartIDUtils, collect, toEthereumAddress, waitForCondition } from '@streamr/utils'
 import { Wallet } from 'ethers'
-import { Broker, createBroker } from '../../../../src/broker'
-import { createClient, createTestStream, formConfig, startBroker } from '../../../utils'
-import { formCoordinationStreamId } from '../../../../src/plugins/operator/formCoordinationStreamId'
 import { cloneDeep, set } from 'lodash'
+import { Broker, createBroker } from '../../../../src/broker'
+import { formCoordinationStreamId } from '../../../../src/plugins/operator/formCoordinationStreamId'
+import { createClient, createTestStream, formConfig, startBroker } from '../../../utils'
 
 const {
     delegate,
@@ -61,7 +61,7 @@ describe('OperatorPlugin', () => {
         const publisher = createClient(fastPrivateKey())
         await stream.grantPermissions({
             permissions: [StreamPermission.PUBLISH],
-            user: await publisher.getAddress()
+            user: await publisher.getUserId()
         })
         const publishTimer = setInterval(async () => {
             await publisher.publish({ id: stream.id }, { foo: 'bar' })