+++ packages/sdk/src/Stream.ts
@@ -18,7 +18,13 @@ import { StreamrClientError } from './StreamrClientError'
 import { StreamRegistry } from './contracts/StreamRegistry'
 import { StreamStorageRegistry } from './contracts/StreamStorageRegistry'
 import { StreamrClientEventEmitter } from './events'
-import { PermissionAssignment, PublicPermissionQuery, UserPermissionQuery } from './permission'
+import { 
+    PermissionAssignment,
+    PublicPermissionQuery,
+    toInternalPermissionAssignment,
+    toInternalPermissionQuery,
+    UserPermissionQuery
+} from './permission'
 import { Resends } from './subscribe/Resends'
 import { Subscriber } from './subscribe/Subscriber'
 import { Subscription, SubscriptionEvents } from './subscribe/Subscription'
@@ -354,10 +360,10 @@ export class Stream {
      * @category Important
      */
     async hasPermission(query: Omit<UserPermissionQuery, 'streamId'> | Omit<PublicPermissionQuery, 'streamId'>): Promise<boolean> {
-        return this._streamRegistry.hasPermission({
+        return this._streamRegistry.hasPermission(toInternalPermissionQuery({
             streamId: this.id,
             ...query
-        })
+        }))
     }
 
     /**
@@ -375,7 +381,7 @@ export class Stream {
      * @category Important
      */
     async grantPermissions(...assignments: PermissionAssignment[]): Promise<void> {
-        return this._streamRegistry.grantPermissions(this.id, ...assignments)
+        return this._streamRegistry.grantPermissions(this.id, ...assignments.map(toInternalPermissionAssignment))
     }
 
     /**
@@ -384,7 +390,7 @@ export class Stream {
      * @category Important
      */
     async revokePermissions(...assignments: PermissionAssignment[]): Promise<void> {
-        return this._streamRegistry.revokePermissions(this.id, ...assignments)
+        return this._streamRegistry.revokePermissions(this.id, ...assignments.map(toInternalPermissionAssignment))
     }
 
 }