+++ packages/sdk/src/permission.ts
@@ -1,4 +1,4 @@
-import { UserID } from '@streamr/utils'
+import { ChangeFieldType, HexString, toUserId, UserID } from '@streamr/utils'
 import { MaxUint256 } from 'ethers'
 
 export enum StreamPermission {
@@ -12,7 +12,7 @@ export enum StreamPermission {
 export interface UserPermissionQuery {
     streamId: string
     permission: StreamPermission
-    user: string
+    user: HexString
     allowPublic: boolean
 }
 
@@ -24,9 +24,11 @@ export interface PublicPermissionQuery {
 
 export type PermissionQuery = UserPermissionQuery | PublicPermissionQuery
 
+export type InternalPermissionQuery = ChangeFieldType<UserPermissionQuery, 'user', UserID> | PublicPermissionQuery
+
 export interface UserPermissionAssignment {
     permissions: StreamPermission[]
-    user: string
+    user: HexString
 }
 
 export interface PublicPermissionAssignment {
@@ -36,11 +38,13 @@ export interface PublicPermissionAssignment {
 
 export type PermissionAssignment = UserPermissionAssignment | PublicPermissionAssignment
 
+export type InternalPermissionAssignment = ChangeFieldType<UserPermissionAssignment, 'user', UserID> | PublicPermissionAssignment
+
 export const PUBLIC_PERMISSION_ADDRESS = '0x0000000000000000000000000000000000000000'
 
 export type PermissionQueryResult = {
     id: string
-    userAddress: UserID
+    userAddress: string
 } & ChainPermissions
 
 export interface ChainPermissions {
@@ -51,12 +55,24 @@ export interface ChainPermissions {
     canGrant: boolean
 }
 
-export const isPublicPermissionQuery = (query: PermissionQuery): query is PublicPermissionQuery => {
+export const isPublicPermissionQuery = (query: InternalPermissionQuery): query is PublicPermissionQuery => {
     return (query as PublicPermissionQuery).public === true
 }
 
-export const isPublicPermissionAssignment = (query: PermissionAssignment): query is PublicPermissionAssignment => {
-    return (query as PublicPermissionAssignment).public === true
+export const toInternalPermissionQuery = (query: PermissionQuery): InternalPermissionQuery => {
+    return ('user' in query) 
+        ? { ...query, user: toUserId(query.user) }
+        : query
+}
+
+export const isPublicPermissionAssignment = (assignment: InternalPermissionAssignment): assignment is PublicPermissionAssignment => {
+    return (assignment as PublicPermissionAssignment).public === true
+}
+
+export const toInternalPermissionAssignment = (assignment: PermissionAssignment): InternalPermissionAssignment => {
+    return ('user' in assignment) 
+        ? { ...assignment, user: toUserId(assignment.user) }
+        : assignment
 }
 
 export const streamPermissionToSolidityType = (permission: StreamPermission): bigint => {