+++ packages/sdk/test/integration/Subscriber.test.ts
@@ -1,7 +1,6 @@
 import 'reflect-metadata'
 
 import { fastWallet } from '@streamr/test-utils'
-import { toEthereumAddress } from '@streamr/utils'
 import { Wallet } from 'ethers'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
@@ -70,7 +69,7 @@ describe('Subscriber', () => {
                     privateKey: publisherWallet.privateKey
                 }
             })
-            await publisher.addEncryptionKey(groupKey, toEthereumAddress(publisherWallet.address))
+            await publisher.addEncryptionKey(groupKey, publisherWallet.address)
     
             const sub = await subscriber.subscribe(stream.id)
     
@@ -122,7 +121,7 @@ describe('Subscriber', () => {
                     privateKey: publisherWallet.privateKey
                 }
             })
-            await publisher.addEncryptionKey(groupKey, toEthereumAddress(publisherWallet.address))
+            await publisher.addEncryptionKey(groupKey, publisherWallet.address)
     
             const sub = await subscriber.subscribe({ streamId: stream.id, raw: true })