+++ packages/sdk/test/integration/publisher-key-reuse.test.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata'
 
-import { Wallet } from 'ethers'
 import { fastWallet } from '@streamr/test-utils'
 import { collect } from '@streamr/utils'
+import { Wallet } from 'ethers'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
 import { StreamPermission } from '../../src/permission'
@@ -33,7 +33,7 @@ describe('publisher key reuse', () => {
         stream = await publisher.createStream('/path')
         await stream.grantPermissions({
             permissions: [StreamPermission.SUBSCRIBE],
-            user: await subscriber.getAddress()
+            user: await subscriber.getUserId()
         })
     })